refactor(storage): inline object tagging entrypoints
